refactor: New file notificationsWatcher.ts/Code refactoring
relates #248
